requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.18.4
pandas==1.0.4
pkg-resources==0.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
